# docs/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Doc Capture API
  version: 0.1.0
  description: >
    Server API for document/photo capture, processing and storage.
    Auth via JWT bearer token obtained from /auth/login.
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FileTemplate:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        pattern: { type: string, example: "{date}_{place}_{docType}_{counter}" }
        appliesTo: { type: string, enum: [document, photo, both] }
    UploadResult:
      type: object
      properties:
        id: { type: integer }
        generatedName: { type: string }
        type: { type: string, enum: [document, photo] }
        storage:
          type: object
          properties:
            type: { type: string }
            path: { type: string }
security:
  - bearerAuth: []
paths:
...
  /templates:
    get:
      summary: List naming templates (admin)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FileTemplate' }
    post:
      summary: Create naming template (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, pattern]
              properties:
                name: { type: string }
                pattern: { type: string }
                appliesTo: { type: string, enum: [document, photo, both] }
      responses:
        '201': { description: Created }
  /templates/{id}:
    patch:
      summary: Update naming template (admin)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK }
    delete:
      summary: Delete naming template (admin)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK }
  /files:
    get:
      summary: List uploaded file records (admin), with optional filters
      parameters:
        - { name: userId, in: query, schema: { type: integer } }
        - { name: type, in: query, schema: { type: string, enum: [document, photo] } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200': { description: OK }
  /files/upload:
    post:
      summary: Batch upload of documents/photos for processing and storage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [place, docType, files]
              properties:
                place: { type: string }
                docType: { type: string, enum: [document, photo] }
                files:
                  type: array
                  items: { type: string, format: binary }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UploadResult' }